+++ src/app/propose/page.tsx
@@ -1,8 +1,10 @@
+"use client";
+
 import { OracleQueries } from "@/components";
 import { usePage } from "@/hooks";
 
-export default function Verify() {
-  const page = "verify";
+export default function Propose() {
+  const page = "propose";
   usePage(page);
   return <OracleQueries page={page} />;
 }
