+++ src/app/settled/page.tsx
@@ -1,8 +1,10 @@
+"use client";
+
 import { OracleQueries } from "@/components";
 import { usePage } from "@/hooks";
 
-export default function Propose() {
-  const page = "propose";
+export default function Settled() {
+  const page = "settled";
   usePage(page);
   return <OracleQueries page={page} />;
 }
