+++ src/components/Notifications.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useNotificationsContext } from "@/hooks";
 import { Root as Portal } from "@radix-ui/react-portal";
 import { animated, useTransition } from "@react-spring/web";
